requirements.txt
geopy==1.22.0
pandas==1.0.3
geopandas==0.7.0
scipy==1.4.1
requests==2.23.0
mlxtend==0.17.2
utm==0.5.0
seaborn==0.10.1
pyyaml==5.3.1
jupyter==1.0.0
